update code structure and renaming some scripts
