+++ download/sthub/__init__.py
@@ -1,0 +1,3 @@
+from . import sthub_facility
+
+StorageHubFacility = sthub_facility.StorageHubFacility
