+++ download/sthub/__main__.py
@@ -1,3 +1,3 @@
 print('storagehubpythonfacility/__main__.py executed')
-import storagehubfacility
+import sthub_facility
