+++ download/sthub/sthub_input.py
@@ -1,5 +1,5 @@
 from download.interface.iinput import InputStrategy
-from download.src.wd_sthub import WorkingDomainStHub
+from download.sthub.sthub_wd import WorkingDomainStHub
 
 
 class InStHub(InputStrategy):
@@ -23,26 +23,9 @@ class InStHub(InputStrategy):
         return depth
 
     def get_time_range(self, working_domain):
-        """
-        Check if the time range refers to the same month, then return the date with the pattern YYYYMM
-        @param working_domain: dict with time information:
-                time: date in string format: [YYYYMM]
-        @return a list of date with pattern YYYYMM
-        """
         if 'time' not in working_domain:
             raise Exception("Can't read time from workingDomain")
-        time = working_domain['time']
-
-        # extract time as YYYYMM
-        start_time = time[0][0:4] + time[0][5:7]
-        end_time = time[1][0:4] + time[1][5:7]
-
-        # time_list = self.generate_time_list(start_time, end_time, time_freq)
-        # check if only one month has been requested
-        # if start_time != end_time:
-        #     raise Exception("time error, check the time range: " + start_time + " - " + end_time)
-
-        return [start_time, end_time]
+        return working_domain['time']
 
     def get_wd(self, working_domain_dict, product_id):
         """
